+++ api/server/router/image/image_routes.go
@@ -363,11 +363,17 @@ func (ir *imageRouter) getImagesJSON(ctx context.Context, w http.ResponseWriter,
 		manifests = httputils.BoolValue(r, "manifests")
 	}
 
+	var containerCount bool
+	if versions.GreaterThanOrEqualTo(version, "1.47") {
+		containerCount = httputils.BoolValue(r, "container-count")
+	}
+
 	images, err := ir.backend.Images(ctx, imagetypes.ListOptions{
-		All:        httputils.BoolValue(r, "all"),
-		Filters:    imageFilters,
-		SharedSize: sharedSize,
-		Manifests:  manifests,
+		All:            httputils.BoolValue(r, "all"),
+		Filters:        imageFilters,
+		SharedSize:     sharedSize,
+		Manifests:      manifests,
+		ContainerCount: containerCount,
 	})
 	if err != nil {
 		return err