+++ client/image_list.go
@@ -52,8 +52,13 @@ func (cli *Client) ImageList(ctx context.Context, options image.ListOptions) ([]
 	if options.SharedSize && versions.GreaterThanOrEqualTo(cli.version, "1.42") {
 		query.Set("shared-size", "1")
 	}
-	if options.Manifests && versions.GreaterThanOrEqualTo(cli.version, "1.47") {
-		query.Set("manifests", "1")
+	if versions.GreaterThanOrEqualTo(cli.version, "1.47") {
+		if options.Manifests {
+			query.Set("manifests", "1")
+		}
+		if options.ContainerCount {
+			query.Set("container-count", "1")
+		}
 	}
 
 	serverResp, err := cli.get(ctx, "/images/json", query, nil)