+++ docs/api/version-history.md
@@ -17,8 +17,6 @@ keywords: "API, Docker, rcli, REST, documentation"
 
 [Docker Engine API v1.47](https://docs.docker.com/engine/api/v1.47/) documentation
 
-* `Sysctls` in `HostConfig` (top level `--sysctl` settings) for `eth0` are no
-  longer migrated to `DriverOpts`, as described in the changes for v1.46.
 * `POST /networks/create` now has an `EnableIPv4` field. Setting it to `false`
   disables IPv4 IPAM for the network. It can only be set to `false` if the
   daemon has experimental features enabled.
@@ -40,7 +38,7 @@ keywords: "API, Docker, rcli, REST, documentation"
   `net.ipv4.config.eth0.log_martians=1`, use
   `net.ipv4.config.IFNAME.log_martians=1`. In API versions up-to 1.46, top level
   `--sysctl` settings for `eth0` will be migrated to `DriverOpts` when possible. 
-  This automatic migration will be removed for API versions 1.47 and greater.
+  This automatic migration will be removed in a future release.
 * `GET /containers/json` now returns the annotations of containers.
 * `POST /images/{name}/push` now supports a `platform` parameter (JSON encoded
   OCI Platform type) that allows selecting a specific platform manifest from